+++ odoo/local-src/specific_sale/models/__init__.py
@@ -3,3 +3,4 @@ from . import sale
 from . import sale_order_line
 from . import account_invoice
 from . import sale_subscription
+from . import subscription_template